+++ test/UnitTests/SendingRequestTests.cs
@@ -61,7 +61,7 @@ namespace UnitTests
                 //
                 // Act
                 //
-                async () => await mediator.SendAsync<PingResponse>(ping)
+                async() => await mediator.SendAsync<PingResponse>(ping)
             );
 
             Assert.IsType<ReturnFunctionIsNullException>(actualEx.Result);
@@ -93,7 +93,7 @@ namespace UnitTests
                 //
                 // Act
                 //
-                async () => await mediator.SendAsync<PingResponse>(ping, cts.Token)
+                async() => await mediator.SendAsync<PingResponse>(ping, cts.Token)
             );
 
             Assert.IsType<ReturnFunctionIsNullException>(actualEx.Result);
@@ -229,29 +229,10 @@ namespace UnitTests
         }
 
         [Fact]
-        public void Send_Not_Configured_Throws_PipelineNotFoundException()
-        {
-            var services = new ServiceCollection();
-            services.AddFluentMediator(m =>
-            { });
-
-            services.AddScoped<IPingHandler, PingHandler>();
-            var provider = services.BuildServiceProvider();
-            var mediator = provider.GetRequiredService<IMediator>();
-
-            var ping = new PingRequest("Ping");
-            var actualEx = Record.Exception(() => mediator.Send<PingResponse>(ping));
-
-            Assert.NotNull(actualEx);
-            Assert.IsType<PipelineNotFoundException>(actualEx);
-        }
-
-        [Fact]
         public void Send_Throws_Exception_Null_Requests()
         {
             var services = new ServiceCollection();
-            services.AddFluentMediator(m =>
-            { });
+            services.AddFluentMediator(m => { });
 
             services.AddScoped<IPingHandler, PingHandler>();
             var provider = services.BuildServiceProvider();